+++ lib/deployex/deployex.ex
@@ -1,10 +1,10 @@
 defmodule Deployex do
   @moduledoc """
   DeployEx is a lightweight tool designed for managing deployments in Elixir
-  applications without relying on additional deployment tools like Docker or
-  Kubernetes. Its primary goal is to utilize the mix release package for executing
-  full deployments or hot-upgrades, depending on the package's content, while
-  leveraging OTP distribution for monitoring and data extraction.
+  and Gleam applications without relying on additional deployment tools like
+  Docker or Kubernetes. Its primary goal is to utilize the release package
+  for executing full deployments or hot-upgrades, depending on the package's
+  content, while leveraging OTP distribution for monitoring and data extraction
   """
 
   alias Deployex.OpSys